Rearrange a bunch of functions from prepare_tokenizer_treebank to a common file
Move the read/write conllu functions to a common folder so they can be used elsewhere
Move the MWT_RE etc as well
Move prepare_treebank_labels to common (and rename it)
Move convert_conllu_to_txt as well
Refactor a tokenizer_conllu_name function
