+++ stanza/tests/tokenization/test_tokenization_orchid.py
@@ -8,7 +8,7 @@ import xml.etree.ElementTree as ET
 import stanza
 from stanza.tests import *
 
-from stanza.utils.datasets.prepare_tokenizer_treebank import convert_conllu_to_txt
+from stanza.utils.datasets.common import convert_conllu_to_txt
 from stanza.utils.datasets.tokenization.convert_th_orchid import parse_xml
 from stanza.utils.datasets.tokenization.process_thai_tokenization import write_section
